requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ajax==3.0.0
aniso8601==7.0.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
AwesomeTkinter==2021.11.8
backports.zoneinfo==0.2.1
bcrypt==3.2.0
blinker==1.4
botocore==1.29.141
Brlapi==0.7.0
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.2
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
decorators==2.0.7
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dj-rest-auth==3.0.0
Django==4.0.10
django-allauth==0.52.0
django-appconf==1.0.5
django-bootstrap3-datetimepicker==2.3
django-core==1.4.1
django-cors-headers==3.14.0
django-countries==7.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-filter==22.1
django-node==4.0.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-react==0.11.1
django-react-tools==0.2.15
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-seed==0.3.1
django-simple-captcha==0.5.1
django-storages==1.13.2
django-utils-six==2.0
django-webpack==3.1.0
django2-bootstrap3-datetimepicker==3.0.3
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
dnspython==2.1.0
docker-pycreds==0.4.0
duplicity==0.8.12.0
entrypoints==0.4
et-xmlfile==1.1.0
fabio==2022.12.1
Faker==18.9.0
fasteners==0.14.1
FireDM==2022.4.14
frozenlist==1.3.3
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
gpg===1.13.1-unknown
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h5py==3.8.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jmespath==1.0.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==1.1.0
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mutagen==1.46.0
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.2.0
olefile==0.46
openai==0.27.2
openpyxl==3.1.1
optional-django==0.1.0
packaging==21.3
paramiko==2.9.3
pathtools==0.1.2
pexpect==4.8.0
phonenumbers==8.13.8
Pillow==7.0.0
plyer==2.1.0
promise==2.3
protobuf==4.22.1
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.18.0
pycups==1.9.73
pycurl==7.45.2
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==2.7.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypng==0.20220715.0
pyRFC3339==1.1
pystray==0.19.4
python-apt==2.0.1+ubuntu0.20.4.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-decouple==3.8
python-environ==0.4.54
python-xlib==0.33
python3-openid==3.2.0
pytz==2022.1
pyxattr==0.6.1
pyxdg==0.27
PyYAML==5.3.1
qrcode==7.4.2
react==4.3.0
reportlab==3.5.34
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
reverse==0.1.0
Rx==1.6.3
s3transfer==0.3.3
SecretStorage==3.3.1
sentry-sdk==1.17.0
setproctitle==1.3.2
silx==1.1.2
singledispatch==4.0.0
six==1.16.0
smmap==5.0.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
systemd-python==234
toposort==1.10
tqdm==4.65.0
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.2
usb-creator==0.3.7
validate-email==1.3
wadllib==1.3.6
wandb==0.14.0
websockets==11.0.3
xdg==5
xkit==0.0.0
yarl==1.8.2
youtube-dl==2021.12.17
yt-dlp==2023.3.4
zipp==1.0.0
